refactor(frontend): retire the formats the redesign superseded
The Models/Settings redesign left three kinds of residue behind: a row
language nothing speaks any more, primitives that outlived the components
they were extracted for, and tokens and stylesheets for surfaces that no
longer exist. knip reports zero findings because every one of these is
still referenced by something equally dead, so this removes the chains
rather than the leaves.

src/ui/list.tsx is gone. It held a generic row language from when several
pages shared one; those pages now speak the catalog table language, which
left `ListRow` and `RowValue` wrapped one-to-one by `SettingsRow` and
`SettingsValue`, `KeyValueRow` called only by the server status page, and
`EmptySafeNotice` called by nothing at all. Each survivor moves to its one
caller — the settings row markup into settings-ui, the key/value line into
features/logs as StatusKeyValueRow — so a reader finds the markup where it
renders instead of behind two hops of indirection.

src/ui/page-state.tsx is gone. Its single caller invoked it as a plain
function rather than as an element, and its loading branch was already
unreachable because UsagePage returns a skeleton first. Only the error case
survived, so the page now draws that itself with ErrorBox.

model-catalog-card.tsx no longer exports a card — the page is a table now —
so it is model-brand.tsx, and the three helpers only it uses stop being
exported.

The dashboard's local status dot is RunStateDot. It is a deliberately
different mark from ui/status's StatusDot (square, no radius, for the dense
instrument sheet) and should not share its name.

Removed 28 lucide re-exports no file imports, ~40 CSS custom properties
nothing reads — including the whole shadcn/Radix compatibility block, of
which only --ring and --destructive-foreground ever had a consumer — and
the stylesheets for retired surfaces: .message-content (superseded by
.chat-markdown), .typing-dots, .zai-gradient-text, the activity-preview
shimmer, the scroll-shadow helpers, and the .theme-zai-* class hooks that
sat beside the [data-theme] selectors the theme runtime actually sets.

Nothing here changes what any surface renders.
