+++ frontend/src/app/styles/globals/base.css
@@ -179,32 +179,6 @@ a.nav-link:hover,
   color: #fff;
 }
 
-/* Scroll shadows */
-.scroll-shadow-top {
-  box-shadow: inset 0 10px 10px -10px rgba(0, 0, 0, 0.3);
-}
-
-.scroll-shadow-bottom {
-  box-shadow: inset 0 -10px 10px -10px rgba(0, 0, 0, 0.3);
-}
-
-.scroll-shadow-both {
-  box-shadow:
-    inset 0 10px 10px -10px rgba(0, 0, 0, 0.3),
-    inset 0 -10px 10px -10px rgba(0, 0, 0, 0.3);
-}
-
-.text-fade-right::after {
-  content: "";
-  position: absolute;
-  right: 0;
-  top: 0;
-  bottom: 0;
-  width: 40px;
-  background: linear-gradient(to right, transparent, var(--surface));
-  pointer-events: none;
-}
-
 /* Left sidebar scroller: no scrollbar at rest, a slim overlay thumb only while
    the pointer is over the sidebar. The 6px gutter is stable so rows do not
    shift when the thumb appears. */
