+++ frontend/src/ui/icon-registry.ts
@@ -50,7 +50,6 @@ export {
 // Lucide-react re-exports — all icons used across the app
 export {
   Activity,
-  AudioLines,
   AlertCircle,
   AlertTriangle,
   Archive,
@@ -58,9 +57,7 @@ export {
   ArrowUp,
   ArrowUpCircle,
   AtSign,
-  Bell,
   Boxes,
-  Braces,
   Brain,
   Cable,
   Check,
@@ -68,15 +65,10 @@ export {
   ChevronDown,
   ChevronLeft,
   ChevronRight,
-  ChevronUp,
   CornerDownRight,
-  CircleAlert,
   CircleDot,
   Clock,
   Code,
-  Code2,
-  Columns2,
-  Compass,
   Copy,
   Cpu,
   Database,
@@ -86,9 +78,7 @@ export {
   Eye,
   EyeOff,
   File,
-  FileCode2,
   FilePenLine,
-  FileSearch,
   FileText,
   Filter,
   FolderTree,
@@ -98,72 +88,54 @@ export {
   Globe,
   Globe2,
   GraduationCap,
-  HardDrive,
   Heart,
   Info,
   Laptop,
   Layers,
-  LayoutDashboard,
   Keyboard,
   KeyRound,
   Link,
   ListChecks,
-  Circle,
   Loader2,
-  LoaderCircle,
   LogOut,
   Menu,
-  MessageCircle,
   MessageSquare,
   MessageSquarePlus,
-  Microchip,
   Minus,
   Monitor,
   Moon,
   MoreVertical,
   Network,
   Paintbrush,
-  Paperclip,
-  PanelLeftClose,
-  PanelLeftOpen,
   PanelRight,
-  PanelRightClose,
-  PanelRightOpen,
   Pause,
   Pin,
-  PinOff,
   Play,
   Plus,
   RefreshCw,
   Rocket,
   RotateCcw,
-  Rows2,
   Save,
   Search,
   Server,
   ServerCog,
   Settings,
-  ShieldCheck,
   Slash,
   Smartphone,
   Sparkles,
   Square,
   SquarePen,
   Sun,
-  Tags,
   Target,
   Terminal,
   TerminalSquare,
-  TrendingUp,
   TriangleAlert,
   Trash2,
   Upload,
   Variable,
-  Volume2,
   Wrench,
   X,
   XCircle,
   Zap,
   Plug,
-  PlugZap,
 } from "lucide-react";
