+++ frontend/src/ui/index.ts
@@ -36,9 +36,6 @@ export type { CardProps, CardPadding } from "./card";
 export { Alert } from "./alert";
 export type { AlertProps, AlertVariant } from "./alert";
 
-export { PageState } from "./page-state";
-export type { PageStateProps } from "./page-state";
-
 export { RefreshButton } from "./refresh-button";
 export type { RefreshButtonProps } from "./refresh-button";
 
@@ -51,8 +48,6 @@ export {
   TabbedPage,
 } from "./page";
 export type { SectionNavItem, PageWidth } from "./page";
-
-export { ListRow, RowValue, EmptySafeNotice, KeyValueRow } from "./list";
 
 export { Slider } from "./slider";
 export { SegmentedControl } from "./segmented-control";
